refactor(ai): split commands/ai.rs into modules
