+++ src-tauri/src/commands/ai/mock_provider_tests.rs
